backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pillow
python-dateutil
aiofiles
opencv-python
numpy
